requirements.txt
requests
pandas
ollama
pyyaml